+++ checky/checky/View/Setting/ViewModuel/CategoriListCellView.swift
@@ -1,0 +1,64 @@
+//
+//  CategoriListCellView.swift
+//  checky
+//
+//  Created by song on 2022/11/03.
+//
+
+import SwiftUI
+import EventKit
+import Combine
+
+struct CategoriListCellView: View {
+  @State var color: Color = .white
+  @State var emoji: String = ""
+  let category: EKCalendar
+  let buttonAction: () -> ()
+  var colorChangedNotification = NotificationCenter.default
+    .publisher(for: UserDefaults.didChangeNotification)
+  
+  init(category: EKCalendar, buttonAction: @escaping () -> ()) {
+    self.category = category
+    self.buttonAction = buttonAction
+    self.colorChangedNotification = NotificationCenter.default
+      .publisher(for: UserDefaults.didChangeNotification)
+  }
+  
+  var body: some View {
+    Button(action: {
+      buttonAction()
+    }, label: {
+      ZStack {
+        Rectangle()
+          .fill(Color.white)
+          .frame(height: 30)
+          .frame(maxWidth: .infinity)
+
+        HStack {
+          Text(category.title)
+            .foregroundColor(Color.fontDarkBlack)
+            .font(.title3)
+          
+          Spacer()
+          
+          Text(emoji)
+          
+          Rectangle()
+            .fill(color)
+            .frame(width: 10, height: 10)
+        }
+        .padding(.horizontal)
+      }
+    })
+    .onAppear {
+      color = fetchUserDefaultColor(calendarIdentifier: category.calendarIdentifier)
+      emoji = fetchUserDefaultEmoji(calendarIdentifier: category.calendarIdentifier)
+    }
+    .onReceive(colorChangedNotification) { _ in
+      color = fetchUserDefaultColor(calendarIdentifier: category.calendarIdentifier)
+      emoji = fetchUserDefaultEmoji(calendarIdentifier: category.calendarIdentifier)
+
+    }
+  }
+  
+}
